src/requirements.txt
aniso8601==9.0.1
attrs==21.2.0
bcrypt==3.2.0
blinker==1.4
cffi==1.14.6
click==8.0.1
dataclasses
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.9
Flask-JWT-Extended==4.2.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.13.0
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
greenlet==1.1.1
idna==3.2
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
marshmallow==3.13.0
mysql-connector-python==8.0.22
protobuf==3.18.0
pycparser==2.20
PyJWT==2.1.0
PyMySQL==0.9.3
pyrsistent==0.18.0
pytz==2021.1
six==1.16.0
SQLAlchemy==1.3.18
typing-extensions==3.10.0.2
Werkzeug==0.16.0
WTForms==2.3.3
zipp==3.5.0
waitress==1.4.4
requests==2.26.0
pyzmq==18.1.1
